tests/wpt/tests/html/canvas/offscreen/reset/2d.reset.state.shadow_color.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.reset.state.shadow_color</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.reset.state.shadow_color</h1>
<p class="desc">check that the state is reset</p>


<script>
var t = async_test("check that the state is reset");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  const default_value = ctx.shadowColor;

  ctx.shadowColor = '#ff0000';
  _assert(ctx.shadowColor == '#ff0000', "ctx.shadowColor == '#ff0000'");

  ctx.reset();
  _assert(ctx.shadowColor == default_value, "ctx.shadowColor == default_value");
  t.done();

});
</script>
